Add ability to change configured env vars
Add the ability to change the value of configured environmental
variables by giving each variable it's own status bar item that can be
clicked, providing a prompt to change the value, and adding
a command that allows selecting the env var to change.

Also add the option to provide a list of options the env var can be set
to and the option to document configured env vars.