Store env var modifications in environmentVariableCollection
Store changes to the environmental variables in
context.environmentVariableCollection. This makes the changes persistent
across application restarts and window reloads. It also makes a warning
pop up in active terminals noting the environment was changed (though
this is currently a bit buggy).

It is still possible to overwrite the values with the system env vars
on launch.

Sadly functionality of this is limited due to https://github.com/microsoft/vscode/issues/142826